move hardware independent functions to env
The hardwrae dependent part is moved to the crate hermit_entry.
This allows to move the functions to the hardware independent
section.
